refactor: move each demo app to it's own domain
