+++ frame-component/package.json
@@ -5,14 +5,16 @@
   "main": "index.js",
   "private": true,
   "scripts": {
-    "dev": "vite serve demo",
+    "demo": "run-p demo:*",
+    "demo:parent-page": "vite --port 3000 serve demo/parent",
+    "demo:child-page": "vite --port 3001 serve demo/child",
     "prettier": "prettier --write .",
     "prettier:check": "prettier --check .",
     "prebuild": "npm run prettier",
     "build": "tsc --declaration",
     "lint": "eslint .",
     "test": "jest",
-    "prewdio": "wait-on http://localhost:3000/index.html",
+    "prewdio": "wait-on http://localhost:3000/index.html && wait-on http://localhost:3001/index.html",
     "wdio": "wdio run ./e2e-tests/wdio.conf.ts"
   },
   "repository": {
